WIP: outline documentation accordion
