+++ src/components/nodes/boundaryErrorEvent/index.js
@@ -3,6 +3,7 @@ import boundaryEventConfig from '../boundaryEvent';
 import merge from 'lodash/merge';
 import cloneDeep from 'lodash/cloneDeep';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 
 export const id = 'processmaker-modeler-boundary-error-event';
 
@@ -28,6 +29,7 @@ export default merge(cloneDeep(boundaryEventConfig), {
             {},
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
