+++ src/components/nodes/boundaryEscalationEvent/index.js
@@ -4,6 +4,7 @@ import merge from 'lodash/merge';
 import cloneDeep from 'lodash/cloneDeep';
 import interruptingToggleConfig from '../boundaryEvent/interruptingToggleInspector';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 
 export const id = 'processmaker-modeler-boundary-escalation-event';
 
@@ -31,6 +32,7 @@ export default merge(cloneDeep(boundaryEventConfig), {
             interruptingToggleConfig,
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
