+++ src/components/nodes/boundarySignalEvent/index.js
@@ -3,8 +3,9 @@ import merge from 'lodash/merge';
 import cloneDeep from 'lodash/cloneDeep';
 import boundaryEventConfig from '@/components/nodes/boundaryEvent';
 import interruptingToggleConfig from '@/components/nodes/boundaryEvent/interruptingToggleInspector';
-import {signalSelector, default as signalEventDefinition} from '../signalEventDefinition';
+import { signalSelector, default as signalEventDefinition } from '../signalEventDefinition';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 
 export const id = 'processmaker-modeler-boundary-signal-event';
 
@@ -34,6 +35,7 @@ export default merge(cloneDeep(boundaryEventConfig), {
             signalSelector('Signal that will trigger this boundary event'),
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
