+++ src/components/nodes/eventBasedGateway/index.js
@@ -1,5 +1,6 @@
 import component from './eventBasedGateway.vue';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 import defaultNames from '@/components/nodes/gateway/defaultNames';
 
 const id = 'processmaker-modeler-event-based-gateway';
@@ -48,6 +49,7 @@ export default {
             },
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
