+++ src/components/nodes/exclusiveGateway/index.js
@@ -1,6 +1,7 @@
 import component from './exclusiveGateway.vue';
 import nameConfigSettings from '@/components/inspectors/nameConfigSettings';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 import defaultNames from '@/components/nodes/gateway/defaultNames';
 
 const id = 'processmaker-modeler-exclusive-gateway';
@@ -47,6 +48,7 @@ export default {
             },
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
