+++ src/components/nodes/serviceTask/index.js
@@ -1,6 +1,7 @@
 import taskConfig from '@/components/nodes/task';
 import nameConfigSettings from '@/components/inspectors/nameConfigSettings';
 import advancedAccordionConfig from '@/components/inspectors/advancedAccordionConfig';
+import documentationAccordionConfig from '@/components/inspectors/documentationAccordionConfig';
 
 export default {
   ...taskConfig,
@@ -33,6 +34,7 @@ export default {
             },
           ],
         },
+        documentationAccordionConfig,
         advancedAccordionConfig,
       ],
     },
