+++ tests/e2e/specs/Modeler.spec.js
@@ -428,38 +428,6 @@ describe('Modeler', () => {
     });
   });
 
-  it('should add and remove documentation to element', () => {
-    const position = { x: 300, y: 300 };
-    const baseElements = [
-      nodeTypes.startEvent,
-      nodeTypes.intermediateCatchEvent,
-      nodeTypes.endEvent,
-      nodeTypes.task,
-      nodeTypes.exclusiveGateway,
-      nodeTypes.pool,
-      nodeTypes.textAnnotation,
-    ];
-
-    baseElements
-      .forEach(type => {
-        const docString = `${type} doc!`;
-
-        dragFromSourceToDest(type, position);
-        cy.contains('Advanced').click();
-        cy.get('[name="documentation"]').clear().type(docString);
-        assertDownloadedXmlContainsExpected(docString);
-
-        cy.get('[name="documentation"]').clear();
-        assertDownloadedXmlDoesNotContainExpected('bpmn:documentation');
-
-        getElementAtPosition(position, type)
-          .click({ force: true })
-          .then($element => {
-            getCrownButtonForElement($element, 'delete-button').click({ force: true });
-          });
-      });
-  });
-
   it('after collapsing panels, show inspector panel when element is highlighted', () => {
     cy.get('[data-test="panels-btn"]').click();
     cy.get('[data-test="inspector-container"]').should('not.be.visible');
